Centralise the summarise/export/erase pieces.
